requirements.txt
selenium==4.9.0
webdriver-manager==3.8.6
pandas==2.0.0
numpy==1.24.3
supabase==1.0.3
python-dotenv==1.0.0
openai>=1.0.0
scikit-learn>=1.0.0 